services/llm-service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "LLM microservice for arkstudy"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi",
  "grpcio",
  "grpcio-tools",
  "uvicorn",
  "SQLAlchemy>=2.0",
  "asyncpg",
  "python-dotenv",
  "httpx",
  "tiktoken>=0.7.0",
  "pgvector>=0.3.4",
  "prometheus_fastapi_instrumentator",
  "aiokafka"
]

[tool.uv]
index-url = "https://pypi.org/simple"